Refatoramento de codigo e testes do controler de lista de compras.

+++ src/com/projeto/entidades/ProdutoNaoIndustrializado.java
@@ -55,17 +55,17 @@ public class ProdutoNaoIndustrializado extends Item {
 	@Override
 	public void AtualizarItem(String atributo, String novoValor) {
 		switch (atributo.toLowerCase()) {
-		case "nome":
-			this.setNome(novoValor);
-			break;
-		case "categoria":
-			this.setCategoria(novoValor);
-			break;
-		case "kg":
-			this.setQuilos(Double.parseDouble(novoValor));
-			break;
-		default:
-			throw new IllegalArgumentException("Erro na atualizacao de item: atributo nao existe.");
+			case "nome":
+				this.setNome(novoValor);
+				break;
+			case "categoria":
+				this.setCategoria(novoValor);
+				break;
+			case "kg":
+				this.setQuilos(Double.parseDouble(novoValor));
+				break;
+			default:
+				throw new IllegalArgumentException("Erro na atualizacao de item: atributo nao existe.");
 		}
 	}
 
@@ -102,7 +102,7 @@ public class ProdutoNaoIndustrializado extends Item {
 	 */
 	@Override
 	public String exibirEmLista(int quant) {
-		return ((int) quant) + " " + this.getNome() + ", " + this.getCategoria();
+		return quant + " " + this.getNome() + ", " + this.getCategoria();
 	}
 
 	/**
