+++ src/com/projeto/entidades/ProdutoPorUnidade.java
@@ -53,17 +53,17 @@ public class ProdutoPorUnidade extends Item {
 	@Override
 	public void AtualizarItem(String atributo, String novoValor) {
 		switch (atributo.toLowerCase()) {
-		case "nome":
-			this.setNome(novoValor);
-			break;
-		case "categoria":
-			this.setCategoria(novoValor);
-			break;
-		case "unidade":
-			this.setUnidade(Integer.parseInt(novoValor));
-			break;
-		default:
-			throw new IllegalArgumentException("Erro na atualizacao de item: atributo nao existe.");
+			case "nome":
+				this.setNome(novoValor);
+				break;
+			case "categoria":
+				this.setCategoria(novoValor);
+				break;
+			case "unidade":
+				this.setUnidade(Integer.parseInt(novoValor));
+				break;
+			default:
+				throw new IllegalArgumentException("Erro na atualizacao de item: atributo nao existe.");
 		}
 	}
 	
@@ -97,7 +97,7 @@ public class ProdutoPorUnidade extends Item {
 	 */
 	@Override
 	public String exibirEmLista(int quant) {
-		return ((int)quant) + " " + this.getNome() + ", " + this.getCategoria();
+		return quant + " " + this.getNome() + ", " + this.getCategoria();
 	}
 	
 	/**
